delete app-settings-layout page, consolidate data under content/patterns/app-settings-layout.tsx, refactor patterns/[...slug].tsx to also support single variant patterns
