+++ polaris.shopify.com/content/patterns/app-settings-layout.md
@@ -4,4 +4,6 @@ description: Makes it easy for merchants to scan and find setting groups.
 url: /patterns/app-settings-layout
 previewImg: /images/patterns/pattern-thumbnail-app-settings.png
 order: 10
+githubDiscussionsLink: https://github.com/Shopify/polaris/discussions/7852
+contentFile: 'app-settings-layout.ts'
 ---
