--- polaris.shopify.com/pages/patterns/app-settings-layout.tsx
@@ -1,4 +1,0 @@
-import PatternsAppSettingsPage from '../../src/components/PatternsAppSettingsPage';
-export default function AppSettingsLayoutPage() {
-  return <PatternsAppSettingsPage />;
-}
