--- polaris.shopify.com/src/components/PatternsAppSettingsPage/index.tsx
@@ -1,2 +1,0 @@
-import PatternsAppSettingsPage from './PatternsAppSettingsPage';
-export default PatternsAppSettingsPage;
